+++ commands/games/_help.js
@@ -7,7 +7,7 @@ module.exports = {
      * A help function
      * @param {Object} game The specific game information
      */
-    run: (e, vars, settings, game) => {
+    run: (e, vars, settings) => {
         try {
             var games = fs.readdirSync('./commands/games/', {encoding: 'utf-8'})
         } catch (err) {
@@ -21,23 +21,13 @@ module.exports = {
             gameList.push(require('./' + game))
         }
 
-        if(!game) {
-            if(!gameList.length) return e.reply("There is no game for now...").then(msg => vars.setEndMessage(msg, "🃏"))
+        if(!gameList.length) return e.reply("There is no game for now...").then(msg => vars.setEndMessage(msg, "🃏"))
 
-            const embed = vars.newEmbed()
-            .setTitle("This is the available games:")
-            for(let game of gameList) {
-                embed.addField(game.name + (game.premium ? " (`" + game.premium + "`)" : ""), game.description + "\n*Start it by using `" + settings.prefix + "game " + game.name + "`*")
-            }
-            e.channel.send(embed)
-        }else {
-
-            const embed = vars.newEmbed()
-            .setTitle("How to play `" + game.name + "` ?")
-            .setDescription(game.help)
-
-            e.channel.send(embed)
-
+        const embed = vars.newEmbed()
+        .setTitle("This is the available games:")
+        for(let game of gameList) {
+            embed.addField(game.name + (game.premium ? " (`" + game.premium + "`)" : ""), game.description + "\n*Start it by using `" + settings.prefix + "game " + game.name + "`*")
         }
+        e.channel.send(embed)
     }
 }